requirements.txt
retry2==0.9.5
aiohttp==3.8.3